Put nix stuff in a subfolder
When I use JJ workspaces I don't have a .git folder in the workspace.
This makes flake.nix misbehave - I copy the whole directory contents in
the /nix/store including the ./target directory which is 1.1GB. If this
were a Git repo I would have copied only the tracked files.

The restructuring fixes this issue.